layout/generic/test/test_bug1408607.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1408607
-->
<head>
  <title>Test for Bug 1408607</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1408607">Mozilla Bug 1408607</a>
<iframe id="display" src="about:blank"></iframe>
<div id="content" style="display: none">

</div>

<div style="margin-left: 1%; border: 1px solid black;">
  <div style="height: 200px; overflow: auto; transform: translateZ(0);">
    <div style="height: 100px;"></div>
    <a id="x" href="./">Hello</a>
    <div style="height: 500px;"></div>
  </div>
</div

<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

window.addEventListener("load", step1);


function step1()
{
  var rect = document.getElementById("x").getBoundingClientRect();
  
  is(document.elementFromPoint(rect.x + rect.width/2, rect.y + rect.height/2), document.getElementById("x"), "Able to hit transformed object");
  SimpleTest.finish();
}


</script>
</pre>
</body>
</html>
